requirements.txt
pyexiftool==0.4.13
requests
pillow
dotenv
pandas
folium
